+++ examples/data-sources/catalystcenter_network_profile_for_sites_assignments/data-source.tf
@@ -1,0 +1,3 @@
+data "catalystcenter_network_profile_for_sites_assignments" "example" {
+  network_profile_id = "ff19242d-9989-434c-beee-92b16b6b30a3"
+}