+++ examples/resources/catalystcenter_network_profile_for_sites_assignments/resource.tf
@@ -1,0 +1,8 @@
+resource "catalystcenter_network_profile_for_sites_assignments" "example" {
+  network_profile_id = "ff19242d-9989-434c-beee-92b16b6b30a3"
+  items = [
+    {
+      id = "bf6cfb4e-750e-434a-b18b-e84c807ea327"
+    }
+  ]
+}